+++ recipe/main.py
@@ -3,7 +3,7 @@ from recipe.commands import Command
 
 def main():
     cmd = Command.parse()
-    cmd.run()
+    cmd.execute()
 
 
 if __name__ == '__main__':